# backend/app/api/v1/endpoints/data.py
"""
Data endpoints for public and private data operations.
"""
from typing import Any, Dict, List, Optional

from fastapi import APIRouter, Depends, HTTPException, Query, status
from fastapi.responses import JSONResponse
from sqlalchemy.ext.asyncio import AsyncSession

from app.core.security import get_current_active_user
from app.db.session import get_db
from app.schemas.data import (
    DataCreate,
    DataInDB,
    DataResponse,
    DataSearchResult,
    DataType,
    DataUpdate,
    DataValidationResult,
    RepositoryEnum,
)
from app.schemas.token import UserResponse

# Create routers
router = APIRouter()
private_router = APIRouter(dependencies=[Depends(get_current_active_user)])

# Public endpoints


@router.get("", response_model=DataInDB)
async def get_data(
    repository: RepositoryEnum,
    data_id: int = Query(..., description="ID of the data to retrieve"),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """
    Retrieve data by ID.
    """
    # TODO: Implement actual data retrieval
    # This is a placeholder implementation
    if data_id < 1:
        raise HTTPException(
            status_code=status.HTTP_404_NOT_FOUND,
            detail="Data not found",
        )
    
    return {
        "id": data_id,
        "repository": repository.value,
        "data_type": "sample",  # Default type
        "data": {"id": data_id, "name": f"Sample Data {data_id}"},
        "created_at": "2023-01-01T00:00:00",
    }


@router.get("/search", response_model=DataSearchResult)
async def search_data(
    repository: RepositoryEnum,
    query: str = Query("*", description="Search query"),
    page: int = Query(1, ge=1, description="Page number"),
    per_page: int = Query(10, ge=1, le=100, description="Items per page"),
    db: AsyncSession = Depends(get_db),
) -> Any:
    """
    Search for data.
    """
    # TODO: Implement actual search
    # This is a placeholder implementation
    return {
        "total": 1,
        "items": [
            {
                "id": 1,
                "repository": repository.value,
                "data_type": "sample",
                "data": {"id": 1, "name": "Sample Data 1"},
                "created_at": "2023-01-01T00:00:00",
            }
        ],
    }


# Private endpoints


@private_router.post("", response_model=DataInDB, status_code=status.HTTP_201_CREATED)
async def create_data(
    data_in: DataCreate,
    repository: RepositoryEnum,
    db: AsyncSession = Depends(get_db),
    current_user: UserResponse = Depends(get_current_active_user),
) -> Any:
    """
    Create new data.
    """
    # TODO: Implement actual data creation
    # This is a placeholder implementation
    return {
        "id": 1,
        "repository": repository.value,
        "data_type": data_in.data_type.value,
        "data": data_in.data,
        "metadata": data_in.metadata or {},
        "created_by": current_user.id,
        "created_at": "2023-01-01T00:00:00",
    }


@private_router.put("/{data_id}", response_model=DataInDB)
async def update_data(
    data_id: int,
    data_in: DataUpdate,
    repository: RepositoryEnum,
    db: AsyncSession = Depends(get_db),
    current_user: UserResponse = Depends(get_current_active_user),
) -> Any:
    """
    Update existing data.
    """
    # TODO: Implement actual data update
    # This is a placeholder implementation
    return {
        "id": data_id,
        "repository": repository.value,
        "data_type": "sample",  # Should get from existing data
        "data": data_in.data,
        "metadata": data_in.metadata or {},
        "created_by": 1,  # Should get from existing data
        "created_at": "2023-01-01T00:00:00",
        "updated_at": "2023-01-02T00:00:00",
    }


@private_router.delete("/{data_id}", status_code=status.HTTP_204_NO_CONTENT)
async def delete_data(
    data_id: int,
    repository: RepositoryEnum,
    db: AsyncSession = Depends(get_db),
    current_user: UserResponse = Depends(get_current_active_user),
) -> None:
    """
    Delete data.
    """
    # TODO: Implement actual data deletion
    # This is a placeholder implementation
    return None


@private_router.post("/validate", response_model=DataValidationResult)
async def validate_data(
    data_in: DataCreate,
    repository: RepositoryEnum,
    db: AsyncSession = Depends(get_db),
) -> Any:
    """
    Validate data.
    """
    # TODO: Implement actual data validation
    # This is a placeholder implementation
    return {
        "valid": True,
        "data": data_in.data,
    }