# backend/app/api/v1/endpoints/validate.py
"""
Validation endpoints for public and private data.
"""
from typing import Any, Dict, List, Optional

from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.responses import JSONResponse
from sqlalchemy.ext.asyncio import AsyncSession

from app.core.security import get_current_active_user
from app.db.session import get_db
from app.schemas.data import (
    DataCreate,
    DataValidationResult,
    RepositoryEnum,
)
from app.schemas.token import UserResponse

# Create routers
router = APIRouter()
private_router = APIRouter(dependencies=[Depends(get_current_active_user)])

# Public endpoints


@router.post("", response_model=DataValidationResult)
async def validate_data(
    data_in: DataCreate,
    repository: RepositoryEnum,
    db: AsyncSession = Depends(get_db),
) -> Any:
    """
    Validate data against the schema.
    """
    # TODO: Implement actual data validation
    # This is a placeholder implementation
    return {
        "valid": True,
        "data": data_in.data,
    }


# Private endpoints


@private_router.post("/private", response_model=DataValidationResult)
async def validate_private_data(
    data_in: DataCreate,
    repository: RepositoryEnum,
    db: AsyncSession = Depends(get_db),
    current_user: UserResponse = Depends(get_current_active_user),
) -> Any:
    """
    Validate private data against the schema.
    """
    # TODO: Implement actual private data validation
    # This is a placeholder implementation
    return {
        "valid": True,
        "data": data_in.data,
    }