# old-backend/public/v1/openapi.yaml
openapi: 3.0.1
info:
  title: EarthRef.org FIESTA API
  description: |
    This is the documentation for the [OpenAPI](https://swagger.io/specification/) definition of EarthRef.org's FIESTA API.

    [EarthRef.org](https://earthref.org) is a geoscience umbrella website for several data repositories.
    These repositories, unified under the mandate to preserve their respective data and promote scientific collaboration in their fields, are also disparate in their schemata.
    The Framework for Integrated Earth Science and Technology Applications (FIESTA) project is creating the tools to easily customize and manage these geoscience repositories and
    the FIESTA API is a core component of that solution.

    # Authentication

    The API is served over HTTPS thereby making HTTP Basic Authentication secure.
    The credentials are the username and password of an active EarthRef.org user account and provide access to that user's private workspace in any of EarthRef.org FIESTA repositories.
    <br/><br/>

  contact:
    email: webmaster@earthref.org
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: '1.3.0'
servers:
  - url: https://api.earthref.org

components:
  securitySchemes:
    EarthRef.org User Account Credentials:
      type: http
      scheme: basic
  parameters:
    repository:
      name: repository
      in: path
      description: Repository
      required: true
      schema:
        type: string
        enum: [MagIC]
    contributionID:
      name: id
      in: query
      description: Contribution IDs
      required: true
      schema:
        type: integer
        minimum: 1
  responses:
    unauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        Authenticate:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorReponse'
    serverError:
      description: Server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/errorReponse'
    noMatches:
      description: No matches found
      content: {}
  schemas:
    messageResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
      example:
        message: 'Here is a message.'
    errorReponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/messageResponse'
      example:
        errors:
          - message: 'There was an error.'

paths:
  /v1/health-check:
    get:
      tags:
        - System
      summary: Check the health of the API
      operationId: v1HealthCheck
      responses:
        200:
          description: Health check passed
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                properties:
                  message:
                    type: string
                example:
                  message: 'Healthy'
        500:
          description: Health check failed
          content:
            application/json:
              schema:
                type: object
                required:
                  - errors
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                      required:
                        - message
                      properties:
                        message:
                          type: string
                      example:
                        message: 'Health check failed'
  /v1/authenticate:
    get:
      tags:
        - People
      summary: Authenticate a username and password
      operationId: v1Authenticate
      security:
        - EarthRef.org User Account Credentials: []
      responses:
        200:
          description: User Record
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                required:
                  - name
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/download:
    get:
      tags:
        - Public Data
      summary: Download the zip archive of files for one or more public contributions
      operationId: v1PublicDownloadFiles
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: n_max_contributions
          in: query
          description: Maximum number of contributions to download in descending timestamp order
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: only_latest
          in: query
          description: Only download the latest version of a contribution
          schema:
            type: boolean
            default: false
        - name: query
          in: query
          description:
            One or more Elasticsearch query strings to match any data model text field
            (https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html#query-string-syntax)
          schema:
            type: string
        - name: id
          in: query
          description: One or more contribution IDs to match
          schema:
            type: array
            items:
              type: integer
              minimum: 1
        - name: doi
          in: query
          description: One or more reference DOIs to match
          schema:
            type: array
            items:
              type: string
        - name: contributor_name
          in: query
          description: One or more contributor names to match
          schema:
            type: array
            items:
              type: string
        - name: reference_title
          in: query
          description: One or more reference titles to match
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: A zip archive of files for the public contributions
          content:
            application/zip: {}
        204:
          $ref: '#/components/responses/noMatches'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/data:
    get:
      tags:
        - Public Data
      summary: Retrieve the data for a public or shared private contribution
      operationId: v1PublicContributionData
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: id
          in: query
          description: The ID of the contribution to retrieve
          schema:
            type: array
            items:
              type: integer
              minimum: 1
        - name: key
          in: query
          description: The private key for retrieving a shared contribution
          schema:
            type: string
      responses:
        200:
          description: Rows are returned in the format requested by the accept header
          content:
            text/plain: {}
            application/json: {}
            application/vnd.ms-excel: {}
        204:
          $ref: '#/components/responses/noMatches'
        401:
          description: Contribution ID and private key do not match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorReponse'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/search/{table}:
    get:
      tags:
        - Public Data
      summary: Retrieve public data rows
      operationId: v1PublicSearchByTable
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: table
          in: path
          required: true
          description: Table name
          schema:
            type: string
        - name: n_max_rows
          in: query
          description: Max number of rows to return
          schema:
            type: integer
            minimum: 1
            maximum: 10000
        - name: from
          in: query
          description: Starting record number for paging
          schema:
            type: integer
            minimum: 1
        - name: query
          in: query
          description:
            One or more Elasticsearch query strings to match any data model text field
            (https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html)
          schema:
            type: array
            items:
              type: string
        - name: included_columns
          in: query
          description: One or more column names that should be included in the matched rows
          schema:
            type: array
            items:
              type: string
        - name: missing_columns
          in: query
          description: One or more column names that should not be included in the matched rows
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: Rows are returned in the format requested by the accept header
          content:
            text/plain: {}
            application/json: {}
            application/vnd.ms-excel: {}
        204:
          $ref: '#/components/responses/noMatches'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/validate:
    post:
      tags:
        - Public Data
      summary: Validate a contribution
      operationId: v1ContributionValidate
      parameters:
        - $ref: '#/components/parameters/repository'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: files
      responses:
        200:
          description: The validation result was returned.
          content:
            application/json:
              schema:
                type: object
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/private/download:
    get:
      tags:
        - Private Data
      summary: Download the zip archive of the files for a private contribution
      operationId: v1PrivateDownloadFiles
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: id
          in: query
          required: true
          description: The ID of the contribution to download
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: A zip archive of files for the private contribution
          content:
            application/zip: {}
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/private/data:
    get:
      tags:
        - Private Data
      summary: Retrieve the data for a private contribution
      operationId: v1PrivateContributionData
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: id
          in: query
          required: true
          description: The ID of the contribution data to retrieve
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: Rows are returned in the format requested by the accept header
          content:
            text/plain: {}
            application/json: {}
            application/vnd.ms-excel: {}
        204:
          $ref: '#/components/responses/noMatches'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/private/validate:
    put:
      tags:
        - Private Data
      summary: Validate a private contribution
      operationId: v1PrivateContributionValidate
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: id
          in: query
          required: true
          description: The ID of the contribution data to validate
          schema:
            type: integer
            minimum: 1
      responses:
        200:
          description: The validation result was returned.
          content:
            application/json:
              schema:
                type: object
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/private/search/{table}:
    get:
      tags:
        - Private Data
      summary: Retrieve private data rows
      operationId: v1PrivateSearchByTable
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - name: table
          in: path
          description: Table name
          required: true
          schema:
            type: string
        - name: n_max_rows
          in: query
          description: Max number of rows to return
          schema:
            type: integer
            minimum: 1
            maximum: 10000
        - name: from
          in: query
          description: Starting record number for paging
          schema:
            type: integer
            minimum: 1
        - name: query
          in: query
          description:
            One or more Elasticsearch query strings to match any data model text field
            (https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-query-string-query.html)
          schema:
            type: array
            items:
              type: string
      responses:
        200:
          description: Rows are returned in the format requested by the accept header.
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
  /v1/{repository}/private:
    post:
      tags:
        - Private Data
      summary: Create a private contribution
      operationId: v1PrivateCreate
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
      responses:
        201:
          description: The private contribution was created.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                properties:
                  id:
                    type: integer
                    minimum: 1
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
    put:
      tags:
        - Private Data
      summary: Replace a private contribution
      description:
        If the contribution is public, a new private contribution will be created in the user's
        private worksapce and the new contribution ID is returned. Tables in the uploaded files
        are combined and replace the existing contribution tables.
      operationId: v1PrivateUpdate
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - $ref: '#/components/parameters/contributionID'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: files
      responses:
        202:
          description: The file(s) were uploaded to the contribution and are being processed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                properties:
                  id:
                    description: The contribution ID to which the data were replaced.
                    type: integer
                    minimum: 1
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
    patch:
      tags:
        - Private Data
      summary: Append to a private contribution
      description:
        If the contribution is public, a new private contribution will be created in the user's
        private worksapce and the new contribution ID is returned. Tables in the uploaded files
        are combined and append to the existing contribution tables.
      operationId: v1PrivateAppend
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - $ref: '#/components/parameters/contributionID'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: array
                  items:
                    type: files
      responses:
        202:
          description: The file(s) were uploaded to the contribution and are being processed.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - rows_added
                properties:
                  id:
                    description: The contribution ID to which the data were appended.
                    type: integer
                    minimum: 1
                  rows_added:
                    description: The number of rows added across all tables.
                    type: integer
                    minimum: 0
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'
    delete:
      tags:
        - Private Data
      summary: Delete a private contribution
      operationId: v1PrivateDelete
      security:
        - EarthRef.org User Account Credentials: []
      parameters:
        - $ref: '#/components/parameters/repository'
        - $ref: '#/components/parameters/contributionID'
      responses:
        200:
          description: The contribution was deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - rowsDeleted
                properties:
                  rowsDeleted:
                    type: integer
                    minimum: 0
        204:
          $ref: '#/components/responses/noMatches'
        401:
          $ref: '#/components/responses/unauthorizedError'
        500:
          $ref: '#/components/responses/serverError'